Safari-609.1.11.1/LayoutTests/fast/forms/listbox-non-contiguous-keyboard-selection-expected.txt

<select> selection test for multiple but non contiguous selection with keyboard.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


1) Select multiple non-adjacent items with the keyboard
PASS selectionPattern("sl1") is "01010"
PASS successfullyParsed is true

TEST COMPLETE
